backend/requirements.txt
fastapi
uvicorn
langchain-core
langchain-community
langchain-groq
langchain-text-splitters
sentence-transformers
chromadb
pypdf
python-dotenv
pydantic
